API for view doc history

+++ grails-app/services/net/hedtech/banner/finance/requisition/system/FinanceApprovalHistoryService.groovy
@@ -29,6 +29,6 @@ class FinanceApprovalHistoryService extends ServiceBase {
      */
     @Transactional(readOnly = true)
     def findByDocumentNumberAndType( documentCode, documentType ) {
-        return FinanceApprovalHistory.fetchByDocumentCodeAndDocType( documentCode, documentType )
+        return FinanceApprovalHistory.fetchByDocumentCodeAndDocType( documentCode, documentType as long )
     }
 }
